Решаем баг с некорректной записью расхода
